+++ apps/bulldozer-js/src/databases/bulldozer/__fixtures__/bulldozer-db-fixture-utils.ts
@@ -21,7 +21,8 @@
 import { decodeBase64, encodeBase64 } from "@hexclave/shared/dist/utils/bytes";
 import { DatabaseSeq } from "../../index.js";
 import { LowLevelDatabase, LowLevelDatabaseDebugSnapshot, LowLevelKvDump, LowLevelKvStore } from "../../low-level/index.js";
-import { PiledriverObject, declarePiledriverDatabase, isPiledriverHeapObjectSymbol } from "../../piledriver/index.js";
+import { declareBasePiledriverDatabase } from "../../piledriver/implementations/base.js";
+import { isPiledriverHeapObjectSymbol, type PiledriverObject } from "../../piledriver/index.js";
 import { BulldozerDatabase, declareBulldozerDatabase } from "../index.js";
 import { exampleFungibleLedgerMigrations } from "../example-schema.js";
 
@@ -62,7 +63,7 @@ const fixtureLedgerRows: Record<string, PiledriverObject> = {
 // Builds a fresh Bulldozer database (example ledger schema) on top of the given low-level backend,
 // applies the migrations, and runs the deterministic workload.
 export async function buildFixtureBulldozerDatabase(lowLevel: LowLevelDatabase): Promise<BulldozerDatabase> {
-  const db = declareBulldozerDatabase(declarePiledriverDatabase(lowLevel), { migrations: exampleFungibleLedgerMigrations });
+  const db = declareBulldozerDatabase(declareBasePiledriverDatabase(lowLevel), { migrations: exampleFungibleLedgerMigrations });
   await db.applyRemainingMigrations();
   await db.withSnapshotReplicated(async snapshot => {
     for (const [rowIdentifier, rowData] of Object.entries(fixtureLedgerRows)) {
@@ -265,5 +266,5 @@ function declareSeededLowLevelDatabase(dump: BulldozerDbDump): LowLevelDatabase 
 // Restores a Bulldozer database from a whole-database dump, ready to be read (and mutated) by the
 // current code.
 export function restoreBulldozerDatabase(dump: BulldozerDbDump): BulldozerDatabase {
-  return declareBulldozerDatabase(declarePiledriverDatabase(declareSeededLowLevelDatabase(dump)), { migrations: exampleFungibleLedgerMigrations });
+  return declareBulldozerDatabase(declareBasePiledriverDatabase(declareSeededLowLevelDatabase(dump)), { migrations: exampleFungibleLedgerMigrations });
 }
